Move more agg and reduction tests
