+++ python/cudf/cudf/tests/test_groupby.py
@@ -1,7 +1,6 @@
 # Copyright (c) 2018-2025, NVIDIA CORPORATION.
 
 import collections
-import datetime
 import itertools
 import operator
 import string
@@ -691,405 +690,6 @@ def test_groupby_apply_return_reindexed_series(as_index):
     expect = pdf.groupby("key", as_index=as_index).apply(pdf_func, **kwargs)
     got = df.groupby("key", as_index=as_index).apply(gdf_func, **kwargs)
     assert_groupby_results_equal(expect, got)
-
-
-def test_advanced_groupby_levels():
-    pdf = pd.DataFrame({"x": [1, 2, 3], "y": [1, 2, 1], "z": [1, 1, 1]})
-    gdf = cudf.from_pandas(pdf)
-    pdg = pdf.groupby(["x", "y"]).sum()
-    gdg = gdf.groupby(["x", "y"]).sum()
-    assert_groupby_results_equal(pdg, gdg)
-    pdh = pdg.groupby(level=1).sum()
-    gdh = gdg.groupby(level=1).sum()
-    assert_groupby_results_equal(pdh, gdh)
-    pdg = pdf.groupby(["x", "y", "z"]).sum()
-    gdg = gdf.groupby(["x", "y", "z"]).sum()
-    assert_groupby_results_equal(pdg, gdg)
-    pdg = pdf.groupby(["z"]).sum()
-    gdg = gdf.groupby(["z"]).sum()
-    assert_groupby_results_equal(pdg, gdg)
-    pdg = pdf.groupby(["y", "z"]).sum()
-    gdg = gdf.groupby(["y", "z"]).sum()
-    assert_groupby_results_equal(pdg, gdg)
-    pdg = pdf.groupby(["x", "z"]).sum()
-    gdg = gdf.groupby(["x", "z"]).sum()
-    assert_groupby_results_equal(pdg, gdg)
-    pdg = pdf.groupby(["y"]).sum()
-    gdg = gdf.groupby(["y"]).sum()
-    assert_groupby_results_equal(pdg, gdg)
-    pdg = pdf.groupby(["x"]).sum()
-    gdg = gdf.groupby(["x"]).sum()
-    assert_groupby_results_equal(pdg, gdg)
-    pdh = pdg.groupby(level=0).sum()
-    gdh = gdg.groupby(level=0).sum()
-    assert_groupby_results_equal(pdh, gdh)
-    pdg = pdf.groupby(["x", "y"]).sum()
-    gdg = gdf.groupby(["x", "y"]).sum()
-    pdh = pdg.groupby(level=[0, 1]).sum()
-    gdh = gdg.groupby(level=[0, 1]).sum()
-    assert_groupby_results_equal(pdh, gdh)
-    pdh = pdg.groupby(level=[1, 0]).sum()
-    gdh = gdg.groupby(level=[1, 0]).sum()
-    assert_groupby_results_equal(pdh, gdh)
-    pdg = pdf.groupby(["x", "y"]).sum()
-    gdg = gdf.groupby(["x", "y"]).sum()
-
-    assert_exceptions_equal(
-        lfunc=pdg.groupby,
-        rfunc=gdg.groupby,
-        lfunc_args_and_kwargs=([], {"level": 2}),
-        rfunc_args_and_kwargs=([], {"level": 2}),
-    )
-
-
-@pytest.mark.parametrize(
-    "func",
-    [
-        lambda df: df.groupby(["x", "y", "z"]).sum(),
-        lambda df: df.groupby(["x", "y"]).sum(),
-        lambda df: df.groupby(["x", "y"]).agg("sum"),
-        lambda df: df.groupby(["y"]).sum(),
-        lambda df: df.groupby(["y"]).agg("sum"),
-        lambda df: df.groupby(["x"]).sum(),
-        lambda df: df.groupby(["x"]).agg("sum"),
-        lambda df: df.groupby(["x", "y"]).z.sum(),
-        lambda df: df.groupby(["x", "y"]).z.agg("sum"),
-    ],
-)
-def test_empty_groupby(func):
-    pdf = pd.DataFrame({"x": [], "y": [], "z": []})
-    gdf = cudf.from_pandas(pdf)
-    assert_groupby_results_equal(func(pdf), func(gdf), check_index_type=False)
-
-
-def test_groupby_unsupported_columns():
-    rng = np.random.default_rng(seed=12)
-    pd_cat = pd.Categorical(
-        pd.Series(rng.choice(["a", "b", 1], 3), dtype="category")
-    )
-    pdf = pd.DataFrame(
-        {
-            "x": [1, 2, 3],
-            "y": ["a", "b", "c"],
-            "z": ["d", "e", "f"],
-            "a": [3, 4, 5],
-        }
-    )
-    pdf["b"] = pd_cat
-    gdf = cudf.from_pandas(pdf)
-    pdg = pdf.groupby("x").sum(numeric_only=True)
-    # cudf does not yet support numeric_only, so our default is False (unlike
-    # pandas, which defaults to inferring and throws a warning about it).
-    gdg = gdf.groupby("x").sum(numeric_only=True)
-    assert_groupby_results_equal(pdg, gdg)
-
-
-def test_list_of_series():
-    pdf = pd.DataFrame({"x": [1, 2, 3], "y": [1, 2, 1]})
-    gdf = cudf.from_pandas(pdf)
-    pdg = pdf.groupby([pdf.x]).y.sum()
-    gdg = gdf.groupby([gdf.x]).y.sum()
-    assert_groupby_results_equal(pdg, gdg)
-    pdg = pdf.groupby([pdf.x, pdf.y]).y.sum()
-    gdg = gdf.groupby([gdf.x, gdf.y]).y.sum()
-    pytest.skip()
-    assert_groupby_results_equal(pdg, gdg)
-
-
-def test_groupby_use_agg_column_as_index():
-    pdf = pd.DataFrame()
-    pdf["a"] = [1, 1, 1, 3, 5]
-    gdf = cudf.DataFrame()
-    gdf["a"] = [1, 1, 1, 3, 5]
-    pdg = pdf.groupby("a").agg({"a": "count"})
-    gdg = gdf.groupby("a").agg({"a": "count"})
-    assert_groupby_results_equal(pdg, gdg, check_dtype=False)
-
-
-def test_groupby_list_then_string():
-    gdf = cudf.DataFrame()
-    gdf["a"] = [0, 1, 0, 1, 2]
-    gdf["b"] = [11, 2, 15, 12, 2]
-    gdf["c"] = [6, 7, 6, 7, 6]
-    pdf = gdf.to_pandas()
-    gdg = gdf.groupby("a", as_index=True).agg(
-        {"b": ["min", "max"], "c": "max"}
-    )
-    pdg = pdf.groupby("a", as_index=True).agg(
-        {"b": ["min", "max"], "c": "max"}
-    )
-    assert_groupby_results_equal(gdg, pdg)
-
-
-def test_groupby_different_unequal_length_column_aggregations():
-    gdf = cudf.DataFrame()
-    gdf["a"] = [0, 1, 0, 1, 2]
-    gdf["b"] = [11, 2, 15, 12, 2]
-    gdf["c"] = [11, 2, 15, 12, 2]
-    pdf = gdf.to_pandas()
-    gdg = gdf.groupby("a", as_index=True).agg(
-        {"b": "min", "c": ["max", "min"]}
-    )
-    pdg = pdf.groupby("a", as_index=True).agg(
-        {"b": "min", "c": ["max", "min"]}
-    )
-    assert_groupby_results_equal(pdg, gdg)
-
-
-def test_groupby_single_var_two_aggs():
-    gdf = cudf.DataFrame()
-    gdf["a"] = [0, 1, 0, 1, 2]
-    gdf["b"] = [11, 2, 15, 12, 2]
-    gdf["c"] = [11, 2, 15, 12, 2]
-    pdf = gdf.to_pandas()
-    gdg = gdf.groupby("a", as_index=True).agg({"b": ["min", "max"]})
-    pdg = pdf.groupby("a", as_index=True).agg({"b": ["min", "max"]})
-    assert_groupby_results_equal(pdg, gdg)
-
-
-def test_groupby_double_var_two_aggs():
-    gdf = cudf.DataFrame()
-    gdf["a"] = [0, 1, 0, 1, 2]
-    gdf["b"] = [11, 2, 15, 12, 2]
-    gdf["c"] = [11, 2, 15, 12, 2]
-    pdf = gdf.to_pandas()
-    gdg = gdf.groupby(["a", "b"], as_index=True).agg({"c": ["min", "max"]})
-    pdg = pdf.groupby(["a", "b"], as_index=True).agg({"c": ["min", "max"]})
-    assert_groupby_results_equal(pdg, gdg)
-
-
-def test_groupby_apply_basic_agg_single_column():
-    gdf = DataFrame()
-    gdf["key"] = [0, 0, 1, 1, 2, 2, 0]
-    gdf["val"] = [0, 1, 2, 3, 4, 5, 6]
-    gdf["mult"] = gdf["key"] * gdf["val"]
-    pdf = gdf.to_pandas()
-
-    gdg = gdf.groupby(["key", "val"]).mult.sum()
-    pdg = pdf.groupby(["key", "val"]).mult.sum()
-    assert_groupby_results_equal(pdg, gdg)
-
-
-def test_groupby_multi_agg_single_groupby_series():
-    rng = np.random.default_rng(seed=0)
-    pdf = pd.DataFrame(
-        {
-            "x": rng.integers(0, 5, size=10000),
-            "y": rng.normal(size=10000),
-        }
-    )
-    gdf = cudf.from_pandas(pdf)
-    pdg = pdf.groupby("x").y.agg(["sum", "max"])
-    gdg = gdf.groupby("x").y.agg(["sum", "max"])
-
-    assert_groupby_results_equal(pdg, gdg)
-
-
-def test_groupby_multi_agg_multi_groupby():
-    rng = np.random.default_rng(seed=0)
-    pdf = pd.DataFrame(
-        {
-            "a": rng.integers(0, 5, 10),
-            "b": rng.integers(0, 5, 10),
-            "c": rng.integers(0, 5, 10),
-            "d": rng.integers(0, 5, 10),
-        }
-    )
-    gdf = cudf.from_pandas(pdf)
-    pdg = pdf.groupby(["a", "b"]).agg(["sum", "max"])
-    gdg = gdf.groupby(["a", "b"]).agg(["sum", "max"])
-    assert_groupby_results_equal(pdg, gdg)
-
-
-def test_groupby_datetime_multi_agg_multi_groupby():
-    rng = np.random.default_rng(seed=0)
-    pdf = pd.DataFrame(
-        {
-            "a": pd.date_range(
-                datetime.datetime.now(),
-                datetime.datetime.now() + datetime.timedelta(9),
-                freq="D",
-            ),
-            "b": rng.integers(0, 5, 10),
-            "c": rng.integers(0, 5, 10),
-            "d": rng.integers(0, 5, 10),
-        }
-    )
-    gdf = cudf.from_pandas(pdf)
-    pdg = pdf.groupby(["a", "b"]).agg(["sum", "max"])
-    gdg = gdf.groupby(["a", "b"]).agg(["sum", "max"])
-
-    assert_groupby_results_equal(pdg, gdg)
-
-
-@pytest.mark.parametrize(
-    "agg",
-    [
-        ["min", "max", "count", "mean"],
-        ["mean", "var", "std"],
-        ["count", "mean", "var", "std"],
-    ],
-)
-def test_groupby_multi_agg_hash_groupby(agg):
-    coll_dict = {letter: float for letter in string.ascii_lowercase}
-    coll_dict["id"] = int
-    gdf = cudf.datasets.timeseries(
-        start="2000",
-        end="2000-01-2",
-        dtypes=coll_dict,
-        freq="1s",
-        seed=1,
-    ).reset_index(drop=True)
-    pdf = gdf.to_pandas()
-    check_dtype = "count" not in agg
-    pdg = pdf.groupby("id").agg(agg)
-    gdg = gdf.groupby("id").agg(agg)
-    assert_groupby_results_equal(pdg, gdg, check_dtype=check_dtype)
-
-
-@pytest.mark.skipif(
-    PANDAS_VERSION < PANDAS_CURRENT_SUPPORTED_VERSION,
-    reason="previous verion of pandas throws a warning",
-)
-@pytest.mark.parametrize(
-    "agg", ["min", "max", "idxmax", "idxmin", "sum", "prod", "count", "mean"]
-)
-def test_groupby_nulls_basic(agg):
-    check_dtype = agg not in _index_type_aggs
-
-    pdf = pd.DataFrame({"a": [0, 0, 1, 1, 2, 2], "b": [1, 2, 1, 2, 1, None]})
-    gdf = cudf.from_pandas(pdf)
-    assert_groupby_results_equal(
-        getattr(pdf.groupby("a"), agg)(),
-        getattr(gdf.groupby("a"), agg)(),
-        check_dtype=check_dtype,
-    )
-
-    pdf = pd.DataFrame(
-        {
-            "a": [0, 0, 1, 1, 2, 2],
-            "b": [1, 2, 1, 2, 1, None],
-            "c": [1, 2, 1, None, 1, 2],
-        }
-    )
-    gdf = cudf.from_pandas(pdf)
-    assert_groupby_results_equal(
-        getattr(pdf.groupby("a"), agg)(),
-        getattr(gdf.groupby("a"), agg)(),
-        check_dtype=check_dtype,
-    )
-
-    pdf = pd.DataFrame(
-        {
-            "a": [0, 0, 1, 1, 2, 2],
-            "b": [1, 2, 1, 2, 1, None],
-            "c": [1, 2, None, None, 1, 2],
-        }
-    )
-    gdf = cudf.from_pandas(pdf)
-
-    # TODO: fillna() used here since we don't follow
-    # Pandas' null semantics. Should we change it?
-
-    assert_groupby_results_equal(
-        getattr(pdf.groupby("a"), agg)().fillna(0),
-        getattr(gdf.groupby("a"), agg)().fillna(0 if agg != "prod" else 1),
-        check_dtype=check_dtype,
-    )
-
-
-def test_groupby_nulls_in_index():
-    pdf = pd.DataFrame({"a": [None, 2, 1, 1], "b": [1, 2, 3, 4]})
-    gdf = cudf.from_pandas(pdf)
-
-    assert_groupby_results_equal(
-        pdf.groupby("a").sum(), gdf.groupby("a").sum()
-    )
-
-
-def test_groupby_all_nulls_index():
-    gdf = cudf.DataFrame(
-        {
-            "a": cudf.Series([None, None, None, None], dtype="object"),
-            "b": [1, 2, 3, 4],
-        }
-    )
-    pdf = gdf.to_pandas()
-    assert_groupby_results_equal(
-        pdf.groupby("a").sum(), gdf.groupby("a").sum()
-    )
-
-    gdf = cudf.DataFrame(
-        {"a": cudf.Series([np.nan, np.nan, np.nan, np.nan]), "b": [1, 2, 3, 4]}
-    )
-    pdf = gdf.to_pandas()
-    assert_groupby_results_equal(
-        pdf.groupby("a").sum(), gdf.groupby("a").sum()
-    )
-
-
-@pytest.mark.parametrize("sort", [True, False])
-def test_groupby_sort(sort):
-    pdf = pd.DataFrame({"a": [2, 2, 1, 1], "b": [1, 2, 3, 4]})
-    gdf = cudf.from_pandas(pdf)
-
-    assert_eq(
-        pdf.groupby("a", sort=sort).sum(),
-        gdf.groupby("a", sort=sort).sum(),
-        check_like=not sort,
-    )
-
-    pdf = pd.DataFrame(
-        {"c": [-1, 2, 1, 4], "b": [1, 2, 3, 4], "a": [2, 2, 1, 1]}
-    )
-    gdf = cudf.from_pandas(pdf)
-
-    assert_eq(
-        pdf.groupby(["c", "b"], sort=sort).sum(),
-        gdf.groupby(["c", "b"], sort=sort).sum(),
-        check_like=not sort,
-    )
-
-    ps = pd.Series([1, 2, 3, 4, 5, 6, 7, 8], index=[2, 2, 2, 3, 3, 1, 1, 1])
-    gs = cudf.from_pandas(ps)
-
-    assert_eq(
-        ps.groupby(level=0, sort=sort).sum().to_frame(),
-        gs.groupby(level=0, sort=sort).sum().to_frame(),
-        check_like=not sort,
-    )
-
-    ps = pd.Series(
-        [1, 2, 3, 4, 5, 6, 7, 8],
-        index=pd.MultiIndex.from_product([(1, 2), ("a", "b"), (42, 84)]),
-    )
-    gs = cudf.from_pandas(ps)
-
-    assert_eq(
-        ps.groupby(level=0, sort=sort).sum().to_frame(),
-        gs.groupby(level=0, sort=sort).sum().to_frame(),
-        check_like=not sort,
-    )
-
-
-def test_groupby_cat():
-    pdf = pd.DataFrame(
-        {"a": [1, 1, 2], "b": pd.Series(["b", "b", "a"], dtype="category")}
-    )
-    gdf = cudf.from_pandas(pdf)
-    assert_groupby_results_equal(
-        pdf.groupby("a").count(),
-        gdf.groupby("a").count(),
-        check_dtype=False,
-    )
-
-
-def test_groupby_index_type():
-    df = cudf.DataFrame()
-    df["string_col"] = ["a", "b", "c"]
-    df["counts"] = [1, 2, 3]
-    res = df.groupby(by="string_col").counts.sum()
-    assert res.index.dtype == cudf.dtype("object")
 
 
 @pytest.mark.parametrize(
